+++ lib/solargraph/workspace.rb
@@ -106,7 +106,7 @@ module Solargraph
     def would_require? path
       require_paths.each do |rp|
         full = File.join rp, path
-        return true if File.exist?(full) or File.exist?(full << ".rb")
+        return true if File.file?(full) || File.file?(full << ".rb")
       end
       false
     end
@@ -133,6 +133,16 @@ module Solargraph
       @gem_rbs_collection ||= read_rbs_collection_path
     end
 
+    # @return [String, nil]
+    def rbs_collection_config_path
+      @rbs_collection_config_path ||= begin
+        unless directory.empty? || directory == '*'
+          yaml_file = File.join(directory, 'rbs_collection.yaml')
+          yaml_file if File.file?(yaml_file)
+        end
+      end
+    end
+
     # Synchronize the workspace from the provided updater.
     #
     # @param updater [Source::Updater]
@@ -144,6 +154,14 @@ module Solargraph
     # @return [String]
     def command_path
       server['commandPath'] || 'solargraph'
+    end
+
+    # True if the workspace has a root Gemfile.
+    #
+    # @todo Handle projects with custom Bundler/Gemfile setups (see DocMap#gemspecs_required_from_bundler)
+    #
+    def gemfile?
+      directory && File.file?(File.join(directory, 'Gemfile'))
     end
 
     private
@@ -214,7 +232,7 @@ module Solargraph
     def configured_require_paths
       return ['lib'] if directory.empty?
       return [File.join(directory, 'lib')] if config.require_paths.empty?
-      config.require_paths.map{|p| File.join(directory, p)}
+      config.require_paths.map { |p| File.join(directory, p) }
     end
 
     # @return [void]
@@ -230,10 +248,11 @@ module Solargraph
 
     # @return [String, nil]
     def read_rbs_collection_path
-      yaml_file = File.join(directory, 'rbs_collection.yaml')
-      return unless File.file?(yaml_file)
+      return unless rbs_collection_config_path
 
-      YAML.load_file(yaml_file)&.fetch('path')
+      path = YAML.load_file(rbs_collection_config_path)&.fetch('path')
+      # make fully qualified
+      File.expand_path(path, directory)
     end
   end
 end