+++ lib/solargraph/yardoc.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'open3'
+
 module Solargraph
   # Methods for caching and loading YARD documentation for gems.
   #
@@ -9,15 +11,25 @@ module Solargraph
     # Build and cache a gem's yardoc and return the path. If the cache already
     # exists, do nothing and return the path.
     #
+    # @param yard_plugins [Array<String>] The names of YARD plugins to use.
     # @param gemspec [Gem::Specification]
     # @return [String] The path to the cached yardoc.
-    def cache(gemspec)
-      path = path_for(gemspec)
+    def cache(yard_plugins, gemspec)
+      path = PinCache.yardoc_path gemspec
       return path if cached?(gemspec)
 
       Solargraph.logger.info "Caching yardoc for #{gemspec.name} #{gemspec.version}"
-      Dir.chdir gemspec.gem_dir do
-        `yardoc --db #{path} --no-output --plugin solargraph`
+      cmd = "yardoc --db #{path} --no-output --plugin solargraph"
+      yard_plugins.each { |plugin| cmd << " --plugin #{plugin}" }
+      Solargraph.logger.debug { "Running: #{cmd}" }
+      # @todo set these up to run in parallel
+      #
+      # @sg-ignore RBS gem doesn't reflect that Open3.* also include
+      #   kwopts from Process.spawn()
+      stdout_and_stderr_str, status = Open3.capture2e(cmd, chdir: gemspec.gem_dir)
+      unless status.success?
+        Solargraph.logger.warn { "YARD failed running #{cmd.inspect} in #{gemspec.gem_dir}" }
+        Solargraph.logger.info stdout_and_stderr_str
       end
       path
     end
@@ -26,16 +38,15 @@ module Solargraph
     #
     # @param gemspec [Gem::Specification]
     def cached?(gemspec)
-      yardoc = File.join(path_for(gemspec), 'complete')
+      yardoc = File.join(PinCache.yardoc_path(gemspec), 'complete')
       File.exist?(yardoc)
     end
 
-    # Get the absolute path for a cached gem yardoc.
+    # True if another process is currently building the yardoc cache.
     #
-    # @param gemspec [Gem::Specification]
-    # @return [String]
-    def path_for(gemspec)
-      File.join(Solargraph::Cache.base_dir, "yard-#{YARD::VERSION}", "#{gemspec.name}-#{gemspec.version}.yardoc")
+    def processing?(gemspec)
+      yardoc = File.join(PinCache.yardoc_path(gemspec), 'processing')
+      File.exist?(yardoc)
     end
 
     # Load a gem's yardoc and return its code objects.
@@ -45,7 +56,7 @@ module Solargraph
     # @param gemspec [Gem::Specification]
     # @return [Array<YARD::CodeObjects::Base>]
     def load!(gemspec)
-      YARD::Registry.load! path_for(gemspec)
+      YARD::Registry.load! PinCache.yardoc_path gemspec
       YARD::Registry.all
     end
   end