+++ spec/type_checker/levels/strict_spec.rb
@@ -40,6 +40,7 @@ describe Solargraph::TypeChecker do
       ))
       expect(checker.problems).to be_one
       expect(checker.problems.first.message).to include('Unresolved call')
+      expect(checker.problems.first.message).not_to include('undefined')
     end
 
     it 'reports undefined method calls with defined roots' do
@@ -48,6 +49,7 @@ describe Solargraph::TypeChecker do
       ))
       expect(checker.problems).to be_one
       expect(checker.problems.first.message).to include('Unresolved call')
+      expect(checker.problems.first.message).to include('String')
       expect(checker.problems.first.message).to include('not_a_method')
     end
 
@@ -55,14 +57,11 @@ describe Solargraph::TypeChecker do
       # @todo This test uses kramdown-parser-gfm because it's a gem dependency known to
       #   lack typed methods. A better test wouldn't depend on the state of
       #   vendored code.
-      gemspec = Gem::Specification.find_by_name('kramdown-parser-gfm')
-      pins = Solargraph::GemPins.build(gemspec)
-      Solargraph::Cache.save('gems', "#{gemspec.name}-#{gemspec.version}.ser", pins)
       source_map = Solargraph::SourceMap.load_string(%(
         require 'kramdown-parser-gfm'
         Kramdown::Parser::GFM.undefined_call
       ), 'test.rb')
-      api_map = Solargraph::ApiMap.new
+      api_map = Solargraph::ApiMap.load_with_cache('.', $stdout)
       api_map.catalog Solargraph::Bench.new(source_maps: [source_map], external_requires: ['kramdown-parser-gfm'])
       checker = Solargraph::TypeChecker.new('test.rb', api_map: api_map, level: :strict)
       expect(checker.problems).to be_empty
@@ -118,6 +117,39 @@ describe Solargraph::TypeChecker do
       expect(checker.problems.first.message).to include('Wrong argument type')
     end
 
+    it 'reports mismatched argument types in chained calls' do
+      checker = type_checker(%(
+        # @param baz [Integer]
+        # @return [String]
+        def bar(baz); "foo"; end
+        bar('string').upcase
+      ))
+      expect(checker.problems).to be_one
+      expect(checker.problems.first.message).to include('Wrong argument type')
+    end
+
+    it 'reports mismatched argument types in calls inside array literals' do
+      checker = type_checker(%(
+        # @param baz [Integer]
+        # @return [String]
+        def bar(baz); "foo"; end
+        [ bar('string') ]
+      ))
+      expect(checker.problems).to be_one
+      expect(checker.problems.first.message).to include('Wrong argument type')
+    end
+
+    it 'reports mismatched argument types in calls inside array literals used in a chain' do
+      checker = type_checker(%(
+        # @param baz [Integer]
+        # @return [String]
+        def bar(baz); "foo"; end
+        [ bar('string') ].compact
+      ))
+      expect(checker.problems).to be_one
+      expect(checker.problems.first.message).to include('Wrong argument type')
+    end
+
     xit 'complains about calling a private method from an illegal place'
 
     xit 'complains about calling a non-existent method'
@@ -129,7 +161,7 @@ describe Solargraph::TypeChecker do
           a[0] = :something
         end
       ))
-      expect(checker.problems.map(&:problems)).to eq(['Wrong argument type'])
+      expect(checker.problems.map(&:message)).to eq(['Wrong argument type'])
     end
 
     it 'complains about dereferencing a non-existent tuple slot'
@@ -940,6 +972,23 @@ describe Solargraph::TypeChecker do
         # @param a [TrueClass]
         def bar(a)
           foo(a)
+        end
+      ))
+      expect(checker.problems.map(&:message)).to eq([])
+    end
+
+    it 'does not complain on defaulted reader with detailed expression' do
+      checker = type_checker(%(
+        class Foo
+          # @return [Integer, nil]
+          def bar
+            @bar ||=
+              if rand
+                 123
+               elsif rand
+                 456
+               end
+          end
         end
       ))
       expect(checker.problems.map(&:message)).to eq([])